reactor: initialize smp::count without a temporary value
before this change, `smp::count` and `nr_cpus` are always initialized
with a certain value respectively, but neither of them are read after
being reassigned with another value. this is confusing and adds the
load of a programmer parsing the source code.

after this change, `nr_cpus` is dropped. and `smp::count` is initialized
at the place where `nr_cpus` is set.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>