build: add checkheaders target
add a "checkheaders" target for validating if headers are
self-contained.

we've been observing build failures due to non-self-contained
headers. the "checkheaders" target should perform the minimal
check to prevent us from using symbols without including the header
which declares it.

this change is inspired by scylladb's "checkheaders" target,
which performs the check by compiling the headers file into objects.
but unlike ninja, CMake refuses to compile header files, we need to
copy them to files with .cc filename extension, and compile the .cc
file instead.

in this change,

1. a new target named "checkheaders-${library}" is created for each
   major Seatar library, like "seastar" and "seastar_testing", for
   checking the headers listed in its `add_library()`
2. and "checkheaders" target will depends on all of these new
   "checkheaders-*" targets, so when we build "checkheaders",
   all of the outdated targets will be build if any of their
   dependencies is updated.
3. the "checkheaders-*" targets' dependencies are just the original
   headers. so if any of these headers is updated, the corresponding
   "checkheaders-*" target is considered outdated. if we run
   "checkheaders" after that, ninja will:
   1. copy the header to a .cc file
   2. compile the generated .cc file

please note, since the dependencies are wired using CMake targets,
so if any of them are updated, ninja will take care of it in a build
tools way. so we don't need to touch a file to force the rebuild.

Fixes #1453
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>