build: use target_link_options() to pass options to linker
instead of using `target_link_libraries()` to pass linker flags,
it is better to use `target_link_options()` for this purpose. As
the former is used for linking against libraries, not for passing
flags.

now that the minimum required CMake version of Seastar is now 3.13,
we have the luxury of using the more appropriate command to do this job.

in this change, `target_link_options()` is used in place of
`target_link_libraries()` when appropriate.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>