resource: transpose the order of checking c.cpus and rc.total_memory
before this change, in allocate(configuration&), `calculate_memory()`
is called before we validate `c.cpus`. but we plan to use `c.cpus`
in `calculate_memory()` so we can reserve memory on a per-shard basis.

the outcome is still the same if we didn't have this change, because
we don't mutate `c.cpus`, when checking for its availability. but
it'd help with the readability if we always do the math based on
verified numbers. so in this change, the code block validating
the availability of cpus is moved before the block calculates the
total memory.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>