prometheus: Don't report labels that starts with __
There's a common practice with Prometheus to use labels that starts with
a double underscore (i.e., __) for internal label manipulation.  This
patch follows the same idea for seastar.

Labels that start with __ will not be reported. This makes it easier to
add labels for metrics manipulation without the cost of sending those
labels on the wire.  Those labels can be used for filtering (like we
already support __name__)

Signed-off-by: Amnon Heiman <amnon@scylladb.com>