metrics: Allows filtering by dynamic labels
The metrics relabel API allows the creation of new labels during run
time.  It would be helpful to add those labels to the set of all
existing labels so that Prometheus will be able to filter by those
labels.

This patch adds the labels to the label collection after performing
metric relabeling.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>