core: s/reserve_additional_memory/reserve_additional_memory_per_shard/
this change is a follow-up of 88b85b88, which added an option named
`reserve_additional_memory` for configuring the memory subsystem, so
we can reserve addition system memory for application which uses
both seastar allocator and system allocator.

but seastar is designed to be a multi-core friendly framework, the
seastar applications always run on multiple shards. if we want to
reserve addition memory for a given application, what we actually
need is to reserve memory for each shard on which the application
allocates memory from system allocator. but in general, user cannot
predict the number of shards of the machine on which the aopplication
is deployed. it is a runtime setting / configuration, and hence can
by retrieved at runtime. so an option for total size of reserve memory
for a sharded application is inherently unscalable.

in this change, the option is renamed to
`reserve_additional_memory_per_shard`.
and when we validate the available memory size,
reserve_additional_memory_per_shard * smp::count is used for the
reserved memory size.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>