docs: drop building-*.md
as none of ubuntu 14.04/15.10/16.04 provides GCC-11 or up
but Seastar requires at least GCC-11 or Clang 14 to build, so
document referencing these releases of distros are updated. but,
Seastar only support the latest two major C++ standards, once the
mainstream C++ compilers, i.e., GCC and Clang supports them. So,
the document referring the exact version of compiler or distros are
likely to be outdated once we bump up the supprted GCC/Clang versions.

also, since the most of the content in building-*.md is repeated
in README.md and HACKING.md, these documents are redundant.

so, in this change,

* doc/building-*.md: instead of bumping up the distro versions,
  building-*.md are dropped
* README.md: some content on how to address OOM kill are added to README.md
* compatibility.md: updated to reference latest GCC versions.

Fixes #1448
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>