semaphore: share state with semaphore_units
The state object is shared between the semaphore
and units extended from it.

It allows the semaphore to be safely moved, destroyed
or reassigned with outstanding units, and allow the
units to function safely after the semaphore is broken
or destroyed, so that the units will not try to
signal the broken/destroyed semaphore.

Fixes #1323

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>