perftune.py: auto-select the same number of IRQ cores on each NUMA
Change the current conservative auto-selection logic that allocates
IRQ cores globally and then distributes them evenly (as much as possible)
among NUMA nodes.

This can result in an odd (in both senses) number of IRQ cores.

For instance on a 2 NUMA node system with 48 cores (96 HTs) without
changing a default value of --irq-core-auto-detection-ratio it is going
to result in 3 IRQ cores auto-selection.

And, as a result, NUMA nodes will be unbalanced since one of them is going
to have more compute CPUs. This can potentially make seastar allocate
RAM from remote NUMA nodes, e.g. if one decides to reduce the default
(7%) amount of reserved (not used by seastar) memory.

This patch makes sure that all NUMA nodes are going to get the same amount of
IRQ cores auto-selected on large multi-NUMA machines which in line is going to
keep compute CPUs balanced too.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>
Message-Id: <20230207023243.1741628-1-vladz@scylladb.com>