build: do not always build seastar as a static library
it hurts the performance if an seastar application links against
a dynamic seastar library, as seastar applications always care
about performance, and seastar is sitting in the critical paths.

but the size of executable also increases significantly if it is
linked against libseastar.a. for instance, the size of library
is 145M with Debug build, and a simple demo application
line_count_demo is 47M if it is linked to libseastar.a, while its
size is 1.2M when linked to libseastar.so.

if we have lots of separated test executables linked agaist seastar
library, it'd save lots of disk space and potentially memory
footprint if they are executed in parallel, as the image of shared
library can be shared in linux OS across executables.

in this change

* an option named "BUILD_SHARED_LIBS" is exposed. and this option
  is OFF by default. so a Seastar application can build Seastar
  as a static library by default as before, or it can build Seastar
  as a dynamic library if it cares about the disk usage or memory
  foot print of the executable image.
* do not pass `-fvisibility=hidden` to as a CXXFLAGS, if
  "BUILD_SHARED_LIBS" is TRUE. so all symbols are exported to
  application linked against libseastar.so. we don't have a set
  of mechinary or policy to expose/version symbols exposed via a
  shared library, so all of them are exposed. in the light of
  C++20 modular support, probably we don't need to introduce this
  mechinary in future.
* disable ASan's `verify_asan_link_order` when "BUILD_SHARED_LIBS"
  is TRUE. otherwise ASan would reference yet-initialized static
  variables via Seastar shared library. see also
  https://github.com/google/sanitizers/wiki/AddressSanitizerFlags

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>