build: build seastar as a shared library in 'debug' or 'dev' mode
in 'debug' and 'dev' mode, we are more likely to be interested in
the fast interation of development, and care less about the results
of benchmark, which should be performed using a "release" build.

to be more spatial efficient either in the sense of disk space usage,
or the memory footprint of multiple instances of seastar applications
when testing in parallel, in this change, Seastar libray is built
as a shared library if the build mode is "debug" or "dev".

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>