thread_pool: prevent missed wakeup when the reactor goes to sleep in parallel with a syscall completion
Seastar issues blocking syscalls in a worker thread. When the syscall
completes, it notifies the reactor thread via a queue. If the reactor
thread is sleeping, or preparing to sleep, it uses an eventfd for the
notification to pull it out of sleep.

In order to prevent races between the reactor falling asleep and the
worker thread waking it up, we have a _main_thread_idle atomic variable.
If it's set, we know we have to use the eventfd for wakeup. However, we
must be sure that the CPU will not reorder loading _main_thread_idle
before the flushing of the completion queue, or the race check in the
reactor thread will fail (in practice, the loading is not ordered before;
instead the preceding writes are held in a private store queue, but the
effect is the same).

In order to prevent the reordering, we used
std::atomic::load(std::memory_order_seq_cst), however this is insufficient.
The documentation states:

"A load operation with this memory order performs an acquire operation, [...]
plus a single total order exists in which all threads observe all
modifications in the same order".

The total order is explained:

"Atomic operations tagged memory_order_seq_cst not only order memory
the same way as release/acquire ordering (everything that happened-before
a store in one thread becomes a visible side effect in the thread that
did a load), but also establish a single total modification order of
all atomic operations that are so tagged."

This does not help, since the write that flushed the completion queue
uses release ordering, not seq_cst.

As a result, we need a stronger barrier. We use std::atomic_thread_fence,
and indeed the generated code (on x86) shows the MFENCE instruction, and
the race [1] is gone.

Since we added the thread fence, we can relax the load memory ordering after
it.

Note, the race is unlikely to cause problems in production. It shows up
when the only thing we wait for is a syscall, but user operations are
unlikely to wait on things like open()/rename()/close(). DMA read/write
and even fsync() are handled by the reactor backend directly. It shows
up only in tests that execute syscalls like rename() intensively, and
have no other work or wakeup source while the rename is happening.

[1] https://github.com/scylladb/scylladb/issues/12795