core/sstring: define <=> operator for sstring
in scylla, we are using <=> in `data_dictionary::storage_options::s3`,
but without the operator<=> defined for sstring, compiler is not able
to implicitly define one for us.

so in this change, an operator<=> is added for sstring.

with this change, Clang-17 should be appeased and to define the <=>
operator for `data_dictionary::storage_options::s3` in Scylla.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>