reactor: syscall thread: wakeup up reactor with finer granularity
If, after we complete a syscall in the syscall thread we find
that the reactor is asleep, we wake it up. Currently, we do that
after we complete all the already-queued syscalls, but this means
that whatever task awaits the first syscall will sleep for too
long - until all syscalls are processed.

Fix this by waking up after every syscall is processed.

This is unlikely to make a difference in practice. First, the reactor
is usually busy. Second, it will spin before falling asleep. Third,
latency-sensitive requests rarely depend on syscalls in Seastar
applications, those are more on the metadata management plane. But
it's better off fixed.

Closes #1507