reactor/scheduling_group: Handle at_destroy queue special in init_new_scheduling_group_key etc (v2)
Fixes #1511

`init_new_scheduling_group_key` iterates all task queues and calls `with_scheduling_group` calls to each.

This in turn will, for all but the active group, call `internal::schedule_in_group`. This will weirdly enough
_work_ even for the _at_destroy group IFF the queue is empty at call, in which case the queue is added to
queues to process. This is of course wrong on oh so many planes.

Now, if the at_destroy queue is _not_ empty, we will instead deadlock here, waiting for a task queue that will
not be scheduled.

This patch
* Special treats the at_destroy queue in init_new_scheduling_group_key
* Changes with_scheduling_group to schedule tasks "checked", which in turn
  ensures we don't try to run things in at_destroy queue ahead of time
    (fatal error, because noexcept)
    * Adds a test case for the issue

v2:
* Fake the "at_exit" group by temporarily assume sched group when calling
  constructors etc. Otherwise atexit would construct in calling sg, which
  breaks scylla.
* Added test for constructor sg uniqueness