seastar-addr2line: extract do_test() and main()
also format the the code to be more PEP8 compliant.

this change is inspired by the intention to use seastar-addr2line
as a python module, so we can use is in yet another python script.
but it turns out we already have a nice module of `addr2line`. so
the change is degraded to a code cleanup.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>