resource: let --memory override --reserve-memory
the help message for `--reserve-memory` reads

> memory reserved to OS (if --memory not specified)

what this help message implies is that, if --memory is specified,
unless the physical memory is not able to provision the necessary
amount of memory, Seastar should support this setting.

but the existing code always takes reserve_memory into consideration
even if `--memory` is specified. this practically prevents user from
running a Seastar application on a machine with memory < 1.5GB, but
> 1GB, with the command line option of `--memory 1G`.

after this change, the memory reserved for OS is only considered
when `--memory` is not set. this should allow user to launch a
Seastar application with the above setting, namely

- physical memory is less than 1.5GB, but greater than 1G
- --memory 1G

See also #1513
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>