build: set the correct version when finding lz4
before this change, `lz4_VERSION` was used for the
detected version. this variable is not set at all.

after this change, `PC_lz4_VERSION` is used instead, so we can
detect the version of lz4, and fail the `find_package()` call, if
the version of installed lz4 does not fulfill the needs.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>