net, websocket: replace Crypto++ calls with GnuTLS
since we already depend on GnuTLS which provides the functions which
we use Crypto++ for. so let's rewrite the Crypto++ function calls
using the ones provided by GnuTLS for less dependencies.

when calculating TCP's Initial Sequence Numbers (ISN for short), we have
two options:

1. have a large-enough buffer, so we can calculate the MD5 hash in a single
   pass. the downside is that we need to copy the 64-byte ISN secretkey
   into the buffer, and the code to fill the "large-enough" buffer would
   be a little bit mess. the upside is that we don't need to allocate
   the hash handle.
2. hash the data with two passes, one using IP addresses and ports, and
   another using the secretkey. the downside is that we have to init and
   deinit a gnutls_hmac_hd_t dynamically instead of reusing the
   registered one.

since we only generate an ISN once for each connection, and the GnuTLS
uses malloc() under the hood for allocating the hash handle struct, also
none of GnuTLS's digest algorithms set its "init" or "deinit" function,
so these are no-op. hence the performance overhead should be negligible.
for the better readability, we use the two-passes approach for
generating the ISN.

please note, GnuTLS does not provide a compile-time constant for the
length of output hash of a given digest algorithm. if it were a C++
library, it could use some templated constexpr function for this purpose
through. so we just use magic numbers for the size of output hash,
and use `assert()` to make sure we don't make mistakes when dupming
these magic numbers from brain (a volatile memory).

the websocket change should be verified by tests/unit/websocket_test
the native tcp change was tested manually, by running httpd with wrk,
like `apps/httpd/httpd --network-stack native`

Fixes #680
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>