prometheus, http: do not expose httpd::* in seastar
before this change, all symbols enclosed in seastar::httpd
namespace are exposed by http/request_parser.hh which is generated
from request_parser.rl, but this is not quite what we intend.

so in this change, `using namespace httpd;` is removed from
`request_parser.rl`, and all places assuming the existence of
symbols exposed by `using namespace httpd;` are updated to
reference `httpd::*`.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>