core: map_reduce: remove unnecessary "mutable"
the `mutable` specifier is used when we need to allow the
lambda body to mutate the variable(s) captured by copy or
to call their non-const member function. but destroying
a captured variable initialized with `std::move(r)` does not
fall into this bucket. so we don't need to add the `mutable`
specifier for the lambda here.

in this change, the `mutable` specifier is dropped.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>