Downgrade expected stall detector warning to info
On most distributions and containers the perf-based stall detector will
fail because /proc/sys/kernel/perf_event_paranoid is set to 2 or higher,
which prevents non-root processes from profiling the kernel.

Instead we log a different message, telling the user what they can do if
they want kernel backtraces, at INFO and only on shard 0 (so we don't
need to see this advice repeated 128 times on a big box).