fstream: drop API level 2 (make_file_output_stream() returning non-future)
Back in 2020, we changed make_file_output_stream() to return a
future<file>. Since this is a breaking change, we introduced
API level 2 to allow applications time to upgrade. More than enough
time has passed, so it is now removed.

Since API level 3 and above is mandatory, the namespace api_v3::and_newer
is also removed, and its contents are incorporated into the seastar
namespace.

The documentation is updated.

I tried to build-test older API levels, but we seem to have
broken them earlier and no one noticed. I plan to follow up
with more API level removals as it's not worth fixing the
compatibility code.

Closes #1556