testing: test_runner: call _exit in thread on error
Seastar doesn't have a clean way to drain background tasks
so as @michoecho explain in
https://github.com/scylladb/seastar/issues/1514#issuecomment-1438615134:
```
while the reactor is being destroyed, some promises are destroyed,
so their futures are marked as broken and the tasks of these futures
are scheduled, but at this point the reactor's task queues might
already be destroyed and the schedule fails with a segfault.
```

Calling _exit after the test completes works around
that, though ideally, we better have a way to stop the reactor
engine cleanly and wait for it to drain all tasks.

We still return 0 on the present path on success
as no seastar tasks should be left behind in this case.

Fixes #1514
Refs scylladb/scylladb#13128

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>