websocket: use gate to guard background tasks.
There are lots of taks running in the background (accept new
connections, process in-coming connections...) in the websocket
server, but not waited properly (e.g, we can listen multiple
addresses, but only the last future created by listen() is waited,
and the future created by connection process is not waited at all),
which is dangerous and can cause coredump.

Here we use gate to wait for all the background tasks.

Besides, do_accepts and do_accept_one are renamed to accept and
accept_one, since methods with do_ prefix are usually internally
used by those without the prefix.

Signed-off-by: Jianyong Chen <baluschch@gmail.com>