core/sstring: do not use ostream_formatter
use fmt::formatter<std::string_view> instead, as fmt::ostream_formatter
creates an extra basic_memory_buffer and a std::basic_streambuf on top
of it, both of this incurs extra overhead. but sstring is used widely
in a Seastar application. so, in this change, let's just reuse
fmt::formatter<std::string_view> for formatting sstring for better
performance.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>