demos: file_demo: pass a string_view to_open_file_dma()
before this change, we pass the filename to `open_file_dma()`
by moving away from a const parameter of the lambda, this catches
a linter's eye. as moving away from a const variable is a no-op.
but what the callee expects is but a string_view. so even if
we can move away from filename, it does not help with the performance
or correctness.

so, in this change, we just change the signature of the lambda in
question, so instead of accepting a `const sstring`, it accepts
a `string_view`. this change has two advantages:

* make it explicit that the lambda relies on the caller to
  keep the filename alive.
* do not send the misleading message that open_file_dma()
  is able to "consume" the given `name` parameter.

Fixes #1549
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>