metrics: use fmt::to_string() when creating a label
before this change, we use `boost::lexical_cast<std::string>()` to
convert a given value to a string when creating a `label_instance`.
but since we are migrating to fmtlib based formatting, it'd be desirable
to format things using fmtlib.

this change has been tested by compiling scylla with the seastar
submodule with the change.

Refs #1544

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>