reactor: trade comment for type constraints
before this change, we are using comments for explaining the typing
requirement of the `Func` template parameter. but the comment cannot
be consumed by compiler, and is not specific enough even from human
reader's perspective.

so, in this change, the "signature: bool()" comments are replaced
with type constraints, so both compiler and human reader can consume
and check this on a more specific common ground.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>