util: backtrace: add print_decoded_backtrace()
After the recent move to use seastar as a shared library in debug
builds, we lost the ability to decode seastar backtraces.

This patch adds a libbacktrace-based helper which prints an
already-decoded backtrace. It might be useful to make decoding
test backtraces easier.

This helper is not production ready yet, because it doesn't check
that features of libbacktrace.h are supported on the platform.