reactor: add an option to print decoded backtraces on abort
Today, to decode a backtrace, we have to download the relevant
binary, run seastar-addr2line on it, and copy-paste the backtrace
from the log to it. This is inconvenient.

It's more convenient to print already-decoded backtraces if possible.
This patch adds an experimental option which prints -- in addition to
the usual backtrace -- a decoded backtrace.

We only do this on abort. It isn't appropriate for other kinds of backtraces
because the decoding might cause huge stalls.