core/on_internal_error: always log error with backtrace
For some reason the backtrace was only logged when
`abort_on_internal_error` was set to true, otherwise the backtrace was
only attached to the exception thrown but not the log message.
This doesn't makes sense and it means that if the exception is not
printed (correctly) the backtrace is lost completely.
This patch fixes this and makes sure the error message is logged with a
backtrace, regardless of what `abort_on_internal_error` is set to.
No performance degradation is expected from this, internal error should
be *very* rare and once they happen, the performance of the affected
code path is not interesting anymore.