test: alien_test: test alien::run_on() also
would be nice if we have more coverage of alien::* helpers. so,
in this change, the test exercising alien::run_on() is added.
please note, noexcept specifier is added to the lambda passed to
alien::run_on() to explain that this function does not throw, also,
this paves the road to the type constraint to be added to
alien::run_on() in a coming commit.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>