fair_queue: Allow more per-tick resource grabbing
Each tick a queue is not allowed to grab all the capacity that's
available on the group. If it hits this local limit, it stops
dispatching, but is allowed to push more data into disk next tick if the
capacity is available. This creates an artificial soft limit on the
local IO queue dispatch rate which, however, can turn into a hard limit
if the CPU is busy and IO queues are polled each latency-goal or more
rarely.

There's an apparent mis-calculation in this per-tick threshold.
Currently the threshold is group_limit / smp::count, which is wrong,
because group limit is shared between queues in the group, not the whole
system. By dividing it by smp::coult the resulting number can happen to
be too low so that the queue is only allowed to push single request per
tick.

The fix is to get the number of queues in the group and make the
threshold be group_limit/group_count instead.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>